det/BDetect/requirements.txt
# scapy
# pyshark
# opencv-python
# numpy
# matplotlib
# pandas
# torch
# torchvision
# flower
# smtplib
# email-validator
# yagmail
# schedule
# psutil
# Pillow
# streamlit-autorefresh
# requests



# Core ML and data processing
torch==2.1.0
numpy==1.26.0
pandas==2.1.1
matplotlib==3.8.0
opencv-python==4.8.1.78
Pillow==10.0.1

# Network traffic monitoring
scapy==2.5.0
pyshark==0.6
netifaces==0.11.0
psutil==5.9.5

# Web and API interactions
requests==2.31.0
flask==2.3.3
python-telegram-bot==20.3

# Scheduling and threading
schedule==1.2.1

# CLI and logging enhancements
tqdm==4.66.1
termcolor==2.3.0
pyfiglet==1.0.2
colorlog==6.7.0

# Optional: for dashboard styling and CSV export
dash==2.14.1
plotly==5.18.0
tabulate==0.9.0

# Optional: for image generation from pcap
seaborn==0.13.0
